requirements.txt
google-genai
python-dotenv
tenacity
requests
beautifulsoup4
pydantic
playwright
Pillow
google-api-core
feedparser
PyYAML
